Refactor MMC5 class in preparation for improving its accuracy, and use limited uint16_t addresses and uint8_t data parameters throughout the library

+++ nes_apu/Nes_Fds_Apu.cpp
@@ -46,9 +46,9 @@ void Nes_Fds_Apu::reset()
 	}
 }
 
-void Nes_Fds_Apu::write_( unsigned addr, int data )
+void Nes_Fds_Apu::write_( uint16_t addr, uint8_t data )
 {
-	unsigned reg = addr - io_addr;
+	uint32_t reg = (uint32_t)addr - io_addr;
 	if ( reg < io_size )
 	{
 		if ( reg < wave_size )
